refactor(KlaviyoCore): move RetryState, flush constants, maxRetries, InvalidField to Core
