+++ Sources/KlaviyoSwift/StateManagement/StateManagement.swift
@@ -17,33 +17,15 @@ import Foundation
 import KlaviyoCore
 import OSLog
 
+// RetryState, FlushConstants (formerly StateManagementConstants flush values),
+// and maxRetries are now in KlaviyoCore/RequestQueueSupport.swift.
+// StateManagementConstants is kept as a thin alias so internal callers
+// continue to compile without churn.
 enum StateManagementConstants {
-    static let cellularFlushInterval = 30.0
-    static let wifiFlushInterval = 10.0
+    static let cellularFlushInterval = FlushConstants.cellularFlushInterval
+    static let wifiFlushInterval = FlushConstants.wifiFlushInterval
     static let maxQueueSize = 200
-    static let initialAttempt = 1
-}
-
-/// Describes how the state machine should handle retrying a request after a failure.
-enum RetryState: Equatable {
-    /// Indicates that the request should be retried immediately (subject to
-    /// the regular flush cadence).
-    ///
-    /// - Parameter currentCount: The attempt number for the *current* request.
-    ///   The value should start at `1` for the very first send and is incremented each
-    ///   time a transient failure (such as a network error) occurs.
-    case retry(_ currentCount: Int)
-
-    /// Indicates that the request should be retried after waiting for a
-    /// server-specified back-off interval. This path is typically triggered by
-    /// an HTTP 429 "Too Many Requests" response that includes a `Retry-After`
-    /// header.
-    ///
-    /// - Parameters:
-    ///   - requestCount: The number of attempts made for this specific request.
-    ///   - totalRetryCount: The total number of attempts made for this request across all retry strategies.
-    ///   - currentBackoff: The remaining time in seconds to wait before the next retry attempt.
-    case retryWithBackoff(requestCount: Int, totalRetryCount: Int, currentBackoff: Int)
+    static let initialAttempt = FlushConstants.initialAttempt
 }
 
 enum KlaviyoAction: Equatable {
